Fix: Update utils folder to align with Ginkgo's conventional structure (#122)
